app/requirements.txt
Flask
pandas
mysql-connector-python
python-dotenv
gunicorn